python/setup.py: don't import its_client
The its_client module is imported only to get the version string.

However, importing the module does not work in cross-compiletion;
indeed, at build time, the dependencies of a python modules are not
available on the build host (they will only be available at runtime
on the target).

Fix that by just duplicating the version string in the setup.py, and by
not importing the module, until we have a better solution.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>