python/main: do not register the signal handler until it is needed
Currently, the signal handler is registered as soon as the main module
is imported, but this is not correct, as the importer may postpone the
actual starting until (much?) later.

So we move the signale regisrtration to until it is actually required,
i.e. just before we start our worker thread.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>