gh-workflow/crates: only publish crates on tags
Currently, we try to publish crates every time a commit in pushed on
master.

However, not every commit on master means that the version of the crates
has changed, or even that the rust code has changed at all in fact. In
such a situation, the workflows are failed on master, because publishing
the crates failed (the version is already on crates.io).

Indeed, a commit can touch all the other stuff in the repository, like
the workflows themselves (like this commit!), or it can change the rust
code without requiring an actual publication of the crates. In the
future, a commit may even touch the soon-to-be-merged implementation in
python...

We change the conditions to publish the crates, based on a version tag
being pushed.

We expect version tag to start with a 'v', like is customary, but this
is still not very robust against arbitrary tags (but they should not be
legion).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>