python/docker: simplify the docker entrypoint script
We now have an entry-point its-client installed by setupttools already,
so we do not need to play tricks with the PYTHONPATH anymore; just call
its-client and be done iwth it.

The exit code was forcefully set to 0, which is totally useless; use the
actual exit code.

Add the now-teaditional "missing line at end of file".

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>