schema/status: add time sources
The type of a time source can be either one of: NTP, PPS, or NMEA; other
types of time sources that may exist in real life are not taken into
account, by lack of acces to such time sources.

Time sources of any type all have:
  * a stratum, which how far it is from the physical clock that drives
    that time source,
  * a state, which is how the time from that source is used,
  * an error, which is an estimatated error of tha ttime source.

Additionally:
  * time sources of type NTP have a host, which is the hostname or IP
    that provides that time source, and their stratum is at least 1;
  * time sources of type PPS have a free-form label, and their stratum
    is always zero (0), as they are actual, physical clocks;
  * time sources of type NMEA have no other field, and their stratum is
    always zero (0), as they too are actual, physical clocks.

To be noted: we define the time sources types with camelCase, as they
are types, not fields in messages, while the time_source field is using
snake_case like all our other fields.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>