github/workflow: print what black would have done
When black decides on a whim that a file does not match its preferred
formatting style, we currently just ask it to just whine, and list the
affected files.

This is not very useful.

First, the list of files is obviously a subset of the files changed in
the branch, and we do not have that many python source files, so the
list is not very useful by itself.

Second, running black locally may very well lead to different results,
because we do not have a golden reference image to use (the test
container is always built on-demand in Github workflows, with un-pinned
version from pip...).

So, this leaves poor souls crying all the tears in their body when they
inadvertently pushed badly formatted code...

We fix that by asking black to also dump the diff it would have applied,
so that at least the issue can be more easily fixed without resorting to
too much hair-pulling when trying to replicate the problem locally...

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>