github/workflows: add workflow to validate schemas
To avoid repeating issues in our schemas, add a workflow that
checks their validity on every push and merge-requests.

We leverage the workflow matrix [0], to spawn as-many jobs as
there are schemas to validate.

However, the matrix in github workflow can only be set by using
expressions [1], and there is only a very limited set of actions
one can do with expressions, and there is nothing that allows
listing a directory to get a set of files.

Instead, we use a preparatory job that generates that listing,
as a json-formatted list, and that list is then injected to fill
in the matrix of the second job.

With much inspiration from [2].

[0] https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs
[1] https://docs.github.com/en/actions/learn-github-actions/expressions
[2] https://tomasvotruba.com/blog/2020/11/16/how-to-make-dynamic-matrix-in-github-actions/

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>