python/its-status: keep timesource data for a little while
In some cases, chronyc might be unable to talk to chronyd (e.g. because
it crashed and/or is being manually restarted to account for a
configuration change.

In this case, the timesources are still valid for a little while,
especially the PPS one, but will ultimately get irrelevant in the near
future, at which point we will not want to report them.

Add a configurable delay during which to keep reporting the last known
time source data.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>