python/its-status: reformat with black
There is no discussion, using black is mandatory. So be it...

All in all, it is pretty trivial and coherent.

Except that, in some situations, we would like to be able to disable
some reformatting, like the list of options passed when calling lshw,
so we can keep related options together on the same line, but black
can't be disengaged locally (like we would do with flake8 for example).

Alas...

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>