python/its-client: use implicit dest of parser.add_argument()
When the first parameter defined is a long option, the name for the
destination variable is automatically derived from the long option.

This property is very commonly used, so it should not be too
surprising to seasoned pythonistas.

This also aligns the variable name for the mqtt broker hostname, to
match the actual option name, but also to match the actual key name
from the configuration file as well. The variable is only ever used
during the configuration step, so we can rename it without firther
side effect.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>