github/workflows/its-client: run on all push
Currently, the its-client workflow that runs the test-suite, only runs
when pushing the maste rbranch, or when opening a PR to the master
branch.

However, we are never supposed to push to he master branch directly, it
always has to happen via a PR, so that would never be triggered.

Furthermore, it is convenient to be ble to push a branch without opening
a PR yet, but just to run the test-suite in the controlled environment.

Change the triggers to run on every push. This will still tigger with
an explicit push to master, should we ever do so (bad!).

Also changeto trigger on PR that target any branch, so as to handle PRs
in maintenance branches (if we ever need one), as well as PRs on
development branches (in case we have different features developped on
various branches and gathered before a final merge to master).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>