python/its-client: don't not autoreconnect
Currently, when the MQTT broker becomes unreachable (e.g. because it
died, or because the cellular network connection dropped), its-client
subsequently dies too.

This is caused by explicitly terminating the MQTT client when an
unexpected disconnection is detected, which causing our main loop to
exit.

In that case, do nothing: paho MQTT is smart eough to kno how to
reconnect on its own. IT may take a bit of time (there is a back-off
mechanism that starts at 1s, and doubles that up to a few tens of
seconds).

When an explicit disconnection is requested (eg. because e received
SIGINT), the main loop will explicitly stop the client, and paho will
not, as expected, try to reconnect, thus gracefully terminating the
whole of its-client, as we want.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>