python/its-client: pass broker info via a dict()
We're going to add another broker, so we'll need to pass as many
variables to the MQTTCLient() constructor, and this is going to become
quite loaded.

Create a dict with the broker configuration, and pass that as a single
parameter to the constructor.

We could have introduced a special type to define an MQTT broker, but
since this is really just static data, a dict() serves the purpose quite
well.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>