python/its-client: be extra aggressive to reconnect to MQTT broker
The default reconnection delay starts at 1s, and doubles up to 120s.

While this recovers relatively quickly from a transient network issue,
the recovery time is way too long when the network dropped for a long
time and eventually comes back up (e.g. after a long tunnel, or an
underground car park...).

Force the reconnect delay to a much shorter maximum, so that we can
recover the MQTT connection faster.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>