python/its-client: add accessor to MQTT client's reception queues
Currently, the ROI object directly accesses the MQTT client's internal
members, which is not very clean object-oriented coding.

Furthermore, we're going to split the queue names so as to make the
prefix configurable, so the members that hold the queue names will
change.

Add an accessor that reurns the name of the requested queue.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>