schemas: restore older versions
It is not because a schema is updated to a new version that all senders
and receivers will automagically and instantly switch to using that
newer version; they may even never be updated, so they will still emit
or consume data that has to conform to an older version of the schema,
so they will still require access to the older versions of a schema to
validate that data.

Furthermore, a receiver may accept, or a sender may generate, more than
one version of the data [0], and may want to be able to validate each
version, so will need to access more than the latest versions of the
schema.

So, we restore all the schema versions that were present in this
repository [1].

However, even with a git repository, it is not trivial to reconstruct
the set of versions for each schema. Indeed, when there is an update, it
is usually done in-place in a previous schema, then at some point the
file got renamed and/or the version field updated, possibly together or
in any order.

This gives many revisions of a schema that refer to the same version.

The heuristic we chose to decide what makes for a schema version, is the
moment the version string is updated in the schema. Whatever changes
were done later on belong to the next version.

This means that we retrieve schemas that may have various issues, like
incorrect references to external entities rather than official ones
(@tigroo31!), or even have validation issues (so we explicitly ignore
them from the validation step).

[0] a sender/receiver pair may negociate the version they want to
exchange, for example.

[1] what existed before the initial commit probably does not matter at
all, and is anyway irremediably lost to the Dark Ages, so ¯\_(ツ)_/¯

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>