python/its-status: switch to using eventfd
Currently, we do all the work in the signal handler. This is very bad.

Offload the work to a loop, which gets triggered from an eventfd. The
eventfd is triggered from the signal handler, which does nothing else.

This paves the way to switch to a better mechanism to get timers
(namely, timerfd).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>