python/its-status: move eventfd in loop()
This will make it easier to switch the implementation of the loop to
something else, and to move the loop out of the main script an into the
module itself, thus hiding the details.

As noticed with a big fat comment, we cheat on the management of our
eventfd: as a file descriptor like any other, it should be closed once it
is no longer needed/used. However, when that is the case, this means the
program is going to exit, either because of an unhandled exception, or
because we explicitly killed it (SIGHUP via Ctrl-C, or SIGTERM when
stopped by systemd).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>