python/its-status: rename thread for GNSS collector
The thread for the GNSS collector is named "gnss", so it is a bit
vague, especially since we are going to introduce more threads,
which will actually be named after the collector they implement.

To avoid confusion, rename the thread to more closely match what it
does. Use the module name as a prefix, and append a string that is
representative of what the thread is, namely: a gpsd client.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>