python/its-status: main does not need configuration items
Now that the eventing is entirely handled in the loop, and that loop
moved out of the main and into the module, there is no need for the main
to handle the configuration items, just to pass them to the loop.

Pass the whole config to loop(), as a keyword parameter, and extract the
required configuration items directly in the loop().

Also pass the configuration as a keyword aprameter to init(), as it
passes it as-is down to the individual plugins, which are ready to cope
with keyword parameters (but only by chance, so we'll fix that later).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>