python/its-status: run all collectors in parallel
Currently, we run all collectors in sequence. However, some collectors,
like the cellular one, can take quite some time to run. This has two
side effects:

  - first, the time to do the whole collect is the sum of the time it
    takes indiviual collectors to run;

  - the first and last collectors do not start at about the same time,
    so the correlation between different data is not that good.

ince we have multi-cores machine, it is possible to have the collectors
run in parallel. Even on uni-processor machines, we'd still gain a bit
of time, as the scheduler could switch when a collector is waiting for
I/O for example.

So, we introduce a kind of light event loop for each of the collectors.
To avoid collectors having to handle that, we introduce a generic loop
that waits for a trigger (in form of an eventfd), calls to the
collector's capture() and collect() methods, stores the data, and
releases the main loop (via another eventfd). For each collector, we
run the corresponding loop in its own thread.

For the trigger eventfd, we need exactly one per collector, but for the
release eventfd, we can just count the number of releases.

As a final note: we do not manage the emitters in threads, because we do
not (yet?) expect those to take too long each: of our two emitters, the
one taking the most time is the MQTT one, and even though paho is the
not the fastest at emitting a message, it is not otherwise overly long;
the other emits on stdout, and that is as fast as it can be.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>