schema: use dots in floating-point notation
We currently use a comma in floating-point notation, but this is purely
french-centric, and is going to somewhat confuse non-french users.
Indeed, people are most likely to expect the english notation, where the
comma is used to separate thousands, while the decimal token is the dot.

Fix our schemas so that the descriptions now all use the english
notation.

Bump the version of the updated schemas accordingly.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>