python/its-status: merge collectors' collect() into capture()
Previously, collector plugins had to provide two functions: one to start
a capture of the data, which could take quite some time, and another
just to provide that data.

The initial idea for that separation was to be able to offload the
capture into threads that would be handled by individual plugins if they
would take too long (like the cellular one) or were asynchronous (like
the gnss one).

It turned out that having each plugin handle its own thread is a bit too
cumbersome, and a generic threading mechanism has been implemented,
whereby the its-status infra handles creating threads for each plugin.

Now, as the per-thread loop now makes obvious, there is no longer any
reason to decorellate the capture and the collect of the data, so we can
eventually merge the two together (which basically amounts to dropping
the collect() function).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>