schema/info: kick up schema update
This is a plain copy of information_schema_1-1-0.json, with just the
version bumped tyo a development one (and a trailig newline added).
(hint: use "git show -C" to check that it is indeed just a copy...)

The version will be eventually bumped to 1.2.0 once the changes in the
schema have been completed.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>