python/its-iqm: extract loading neighbours from file, to a new module
For now, we can only have neighbours defined in a local config file.
But we're supposed to also support loadgin them via http, or even better
in the(far) future, viw MQTT.

Introduce the central module that allows connecting to a central
authority, that can be a file, an http request, or an MQTT connection;
for now, still only file is implemented.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>