python/its-status: make the example .cfg file usable as-is
Currently, the sample defconfig file contains comments at the end of the
assignment lines. The python configparser does not interpretate such
commants, and reports them as part of the value, which causes quite a
mess when loading the file:

    ValueError: invalid literal for int() with base 10: '2947  # the port gpsd listens on'

Fix that by moving the comments above the line they describe.

Reported-by: Frederic GARDES <frederic.gardes@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>