python/its-status: fix API of {cellular,system} collectors
As the configuration is unused by those two collectors, the 'cfg'
parameter was prefixed with an underscore.

However, this is incorrect: it is a keyword-only parameter, so it can't
be underscore-prefixed, or caller would get a TypeError when they try to
instanciate those collectors:

    [...]
      File "/usr/lib/python3.10/site-packages/its_status/__init__.py", line 56, in init
        "obj": mod.Status(*args, **kwargs),
    TypeError: Status.__init__() got an unexpected keyword argument 'cfg'

Fix that by renaming the keyword-only parameter to the non-prefix 'cfg'.

Fixes: 295544c

Note: only positional parameters can be underscore-prefixed to denote
that they are unused.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>