python/its-status: add config option for the data directory
Currently, we hard-code /data as the directory for which to report total
and free space.

However, in some cases, this directory does not exist, and it sould
require being root to create, which makes it less than optimal to test
its-status.

Add a new configuration option to specify that directory. Use / as the
example, as that makes for a ready-to use config file.

Fixes: #63

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>