pytohn/its-client: emergency exit in case of suprious MQTT deconnection
We have a race condition somewhere, whereby the its-client stalls
around a GPS update, a RoI update, MQTT messages published/received,
and a spurious MQTT deconnection.

The root cause for that issue has been eluding us for months now, to
the point where we can only seriously think about doing an emergency
exit when we ever notice the situation, and let the service manager
(e.g. systemd) restart us...

That's ugly, but as stated in #75, its-client is undergoing a complete
rewrite, so we're not going to spend more time on trying to hunt and
fix that issue... :-(

Closes #75.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>