python/its-cient: fix invalid acceleration
When unknown, we currently set the acceleration to 161 (implied: m/s²).
When a CAM message is assembled, that value is multiplied by ten and
rounded to an int (to ETSI dm/s²). This gives an encoded value of 1610.

However, the encoding for the acceleration in the CAM message, defined
by the ETSI acceleration unit, provides for a range [-160...161], with
161 coding an unknown acceleration.

1610 is thus an invalid encoded value, making the whole CAM message
invalid.

The problem is once again a combination of two usual factors:
  - the confusion between values and status, where some magic values
    encode some status rather than having a separate field to report
    the status, and
  - the confusion of units between raw and encoded values.

We fix that by using a non-number, thus invalid, value as the
acceleration, and converting that only at the last moment when we
eventually fill in a CAM. If the value is invalid, we use the magic
value that represent an unknown acceleration (161), otherwise we encode
the acceleration with the proper conversion formula.

Partially closes #76 (invalid 1610 value)

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>