python/its-client: fix computing acceleration
We currently compute the acceleration by dividing the difference between
two positions, by the time elapsed to go from one position to the other.

That's a speed, not an acceleration...

A better approximation for the acceleration, is to actually divide a
difference between the speed at two different points in time, by the
time it took to go from the first speed to the other. (yeah, that's
actually what an acceleration is, assuming constant acceleration in
that time period).

Partially closes #76 (invalid acceleration computation)

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>