schema/status: allow reporting signal metrics for multiple technologies
While there is always going to be a single technology used to establish
a connection, i.e. to to authenticate against the network, there can be
multiple technologies used to transport data; for example, with 5G-NSA,
LTE (aka 4G) is used to authenticate, so that is the tecnology used to
establish the connection, but both LTE and 5G technologies can be used
(simultaneously or alternatingly) to transport data.

Change the schema to denote that, for a single connection, we may have
to report signal KPIs for multiple technologies.

To be noted: it would seem to make sense that we do have at least one
technology to report KPIs for. However, that may even be wrong: to limit
energy usage, we could very well envision that a modem disables all of
its radios for short periods of time, while still having an actve
connection, and temporarily reenables one radio to keep that connection
alive, or when there is actual data to transmit. So we provision for
this case as well; it also makes things simpler to manage.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>