python/its-info: create timer before MQTT client
So far, we create the timer after we create the MQTT client, and most
importantly, after we called client connect_async(). However, we set the
timer value in the on_connect() hook, so we expect the timer to exist
before we can get connected.

So far, this is OK, becaue we do an asynchronous connect, and we only
trigger the actual connection when we later publish our first message.

However, this is a bit fragile, and in case we ever convert to an
immediate connection (not async), there's going to be a race between the
connection and the creation of the timer (the latter almost always
wining this race, but "almost always" is not a guarantee).

Create the timer before we create the client, to ensure this race never
happens in the first place.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>