python/its-client: expect explicit port for mirror broker
Currently, when the mirror broker is configured in a config file, the
MQTT port from the config file is ignored, in favour of the default from
the corresponding argparse option.

Since using command line options is only meant as a debug/testing aid,
users should be expected to explicitly provide the port if they want to
use command line arguments rather than a config file.

So, drop the default from the argparse option, so that the value from
the config file gets used *unless* the user knows better by passing an
explicit value on the command line.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>