python/iqm: fix mqtt authority
The authority.client_id field is optional, so trying to dereference it
could yield a KeyError exception; the current construct, using an 'or',
would not prevent the exception from being raised. Use a try-except
block instead.

The 'message' parameter of the _on_message() callback is actually used,
and is referenced, i.e. without a leading underscore.

The MQTT host and port are not stored as attributes to the object, so
always refer to those from the cfg object, which we do store.

We read the MQTT authority configuration from a ini-style file, so all
fields are strings; the paho MQTT client expects the port to be an int
and does no conversion on its own, so we must cast it ourselves.

Fix the typo in the subscribe() method call.

Those issues have not yet been caught, as the MQTT authority is not
implemented yet, and the PR introducing the IQM only tested the 'file'
method (bad me, my fault).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>