python/iqm: don't over-report loaded neighbours
The 'file' and 'http' methods load the neighbours list from a .cfg
style content, with a section for each neighbour. We use python's
configparser.ConfigParser() to parse those, and there is always a
section named "DEFAULT", thus over-reporting the number of loaded
neighbours by one when getting the len() of the ConfigParser().

Fix that by using the .sections() accessor, as that returns a list
that does not contain the "DEFAULT" section.

To be noted: the 'mqtt' method was not reporting on the number of
loaded neighbours, so add it; it also uses JSON blobs, to conform
with the rest of the payloads that go through the ITS MQTT broker;
that does not have the issue, and we add a comment to dispell any
future possible confusion about any then-perceived discrepancy.

Fixes: #103

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>