python/iqm: allow heterogeneous remote interQueue
The prefix to queues is meant to implement some poor-man's namespaces.
Although the namespace, and thus the prefix, is meant to be common
across the geoservers of a single entity, it is theoretically possible,
and will in practice be quite common, that two geoservers that are
neighbours of each other, will use a different namespaces. This would be
the case across geographical entities, like countries or administrative
regions, or across legal entities, like companies managing different
pay-for-use roads (like french motorways).

Allow neighbours to define their own prefix and suffix.

Closes: #107

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>