python/iqm: make reload mandatory for 'file' and 'http' authorities
The list of neighbours is always going to change over time, so we'll
always have the need to update it. It would be possible to just restart
the IQM when that happens, but that would create a pause in the service,
which is not nice; furthermore, the reloading mechanism is already
implemented.

Make the reload timer mandatory for the http and file authority types.
By design, the MQTT authority has no timer, and the list of neighbours
is updated when we receive it on the specified topic already.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>