python/iqm: abstract the authority type
Currently, the IQM object needs to know what type of Authority is being
used, so that it can instantiate an object of the proper class. This is
a breach in responsibility separation, and should be abstracted away so
that the knowledge about what class implements what type of authority,
is relegated to the authority code.

Introduce a wrapper/demuxer class in the authority module, and that
returns the appropriate object that implements the requested authority
type.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>