python/iqm: provide a dict() to the update_cb() callback
Don't pass a configparser.ConfigParser() object, as it is a little bit
more difficult to handle. Knowing what format the neighbours list is
stored or retrieved as, should only be of concern to each type of
authority client, and should not leak into the queue manager.

The MQTT authority already provides a dict (well, it is an object loaded
from a json blob, so definitely not a configparser.ConfigParser() object
at least).

Update the type annotations for the callback to match this new
semantics.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>