python/iqm: fix anonymous connection to neighbours
We accept that we can connect anonymously to the MQTT broker of the
neighbours, for example if authentication is done with certificates (not
implemented yet, but that's the idea). So we state (badly) that username
and passwords can be set empty to authenticate anonymously.

However, an empty username is semantically not the same as no username.
It is also possible to provide a username without a password, which
again is not the same as an empty password.

Also fix the comment.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>