python/iqm: make the MQTT authority topic configurable
Currently, the topic for the MQTT authority is computed by aggregating
the prefix and the suffix as for the queues, with the queue name
replaced by the hard-coded 'neighbours/' component.

This is not very flexible, and there is no reason the topic for
receiving the neighbours list be formatted like the topic for the
queues.

It also requires that the global configuration item be passed to the
authority object, when it should only be concerned with the authority
configuration.

Make the topic a mandatory config item. If an existing, unlikely setup
was making use of the prefix and suffix, they can just be duplicated in
the authority topic.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>