python/iqm: fix socket-based connection to local MQTT broker
When we connect to the local MQTT broker via a UNIX socket, we have no
host and no port on the config file, so the defaults from the code are
used, None and None. However, when we connect via TCP, we have str and
str (as coming from a configparser.ConfigParser(), which treats
everything as str).

However, the mqtt_client port argument is an int|None, so when we
connect via TCP, we should pass an int, but we pass an str. So far, this
is working, because mqtt_client does an explicit conversion. That
conversion is supperfluous, because the port *is* supposed to be an int
in mqtt_client.

So, we must pass an int or a None when instanciating the local client.
We can blindly convert the config value to an int, as it might be a
None (from the default), so we need to only convert it if we have no
UNIX socket path configured, which implies we are supopsed to have a
valid port number.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>