python/its-vehicle: extend GNSS report
Not all GNSS devices trigger gpsd to emit GST or ATT messages, but when
they do, there are interesting data in there that we want to make
available for enhanced reporting (esp. the acceleration, the error
ellipse...).

Even when those messages are emitted, there is still no guarantee that
any of the fields be present at all, as gpsd explicitly documents that
all fields are optional, so be prepared for that.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>