python/its-quadkeys: return the actually shallower quadkey
To make a shallower quadkey, the new depth was computed so that the
new quadkey would be at least of depth 1, i.e. always with at least
one digit.

However, the new quadkey was not built from that depth, but from the
depth of the original quadkey. That means that, when a quadkey of depth
1 was made shallower, an empty quadkey was returned, which is invalid.

Fix that by using the proper depth.

Fixes: #130.

Reported-by: GARDES Frederic INNOV/IT-S <frederic.gardes@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>