python/its-quadkeys: fix type hint for QuadZone
PEP484 [0] mandates that, for arbitrary argument lists (e.g. *args and
**kwargs), the type for the individual elements whould be hinted, and
while *args is a list (and **wargs a dictionary), it's the items in the
list (the values of the jeys in the dict) that have to be type hinted,
e.g. (from the PEP itself):

    def foo(*args: str, **kwargs: int): ...

specifies that foo takes any number of positional arguments (*args) that
are all str, and any number of keyword arguments that are all int.

Thus, the current type hint for the QuadZone is wrong. What we really
wanted is to be able to pass an arbitrary number of any of those types;
  - a string representing a QuadKey
  - a QuadKey object
  - a list of strings each representing a QuadKey, or of QuadKey objects

So, all the following calls are expected to be valid and equivalent:

  * QuadZone()
  * QuadZone("0", "1", "2")
  * QuadZone(["0", "1", "2"])
  * QuadZone(["0", "1"], "2")
  * QuadZone(["0", "1"], ["2"])
  * QuadZone(QuadKey("0"), QuadKey("1"), QuadKey("2"))
  * QuadZone([QuadKey("0"), QuadKey("1"), QuadKey("2")])
  * QuadZone([QuadKey("0"), QuadKey("1)"], QuadKey("2"))
  * QuadZone([QuadKey("0"), QuadKey("1)"], [QuadKey("2")])
  * QuadZone(QuadKey("0"), ["1", "2"])
  * QuadZone(QuadKey("0"), "1", [QuadKey("2")])
  * and so on...

[0] https://peps.python.org/pep-0484/#arbitrary-argument-lists-and-default-argument-values

Reported-by: GARDES Frederic INNOV/IT-S <frederic.gardes@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>