python/its-quadkeys: fix nothern- or southern-most neighbours
Contrary to what the doc-string suggests, the nother-of() and south_of()
do not return None for the northern- or southern-most QuadKeys. Instead,
it tries to instanciate a QuadKey when there is None, which does fail.

Furthermote, the {north,south}_{east,west}_of() suffer from the same
issue, while they did not document returning None.

Fix those cases, and add a regression test.

Fixes: #147

Reported-by: Hugues OUDEVILLE <hugues.oudeville@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>