python/its-quadkeys: fix QuadZone neighbourhood to the North and South
When looking up neighbours of a QuadZone, we look up neighbours of each
individual QuadKeys. When one such QuadKey is the northern- or
southern-most QuadKey, it has no northern or southern neighbour, but the
named-tuple still have fields for them; those fields are set to None.

This then causes an issue when we are trying to look whether the
individual neighbours are located in the other region, because we did
not filter-out he missing neighbours.

Fix that by explicitly excluding None from the neighbours.

Fixes: #147

Reported-by: Hugues OUDEVILLE <hugues.oudeville@orange.com>
Signed-off-by: Yann E. MORIN <yann.morin@orange.com>