python/its-quadkeys: be a bit more verbose when dumping errors
In case of error, we are dumping the difference between what we expected
and what we got. However, it is often interesting to also have both the
expected and the computed values.

Extend the error reporting as such, and turn all QuadKeys into strings,
as it is easier to review.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>