python/iot3/core: fix MQTT over Unix socket
Although the unix socket transport does not require a TCP port,
paho-mqtt expexts that a valid TCP port be specified when connecting.

Technically, specifying a port is not necessary in that case, so we
could in theory change our call to connect_async() to not pass a port
at al. However, that would make the code a bit more complex.

So, to simplify things, just use a valid TCP port, i.e. one that is
strictly greater than 0. We could have decided to use the default port,
1883, but that would be a bit misleading. Instead, we just use 1, which
is odd enough that it will at least raise an eyebrow to an unsuspecting
reviewer.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>