github/workflows: don't upload python artefacts
We were until now using the GitHub Action actions/upload-artifact@v2 to
upload the python jobs artefacts. However, commit 42f406b7a3c5 (github:
add a python workflow) did not explain why that action was used;
presumably, that was done so as to make the generated python package
available for later download (via the web UI for example).

However, actions/upload-artefact is not meant for releasing artefacts;
instead, it is meant so as to pass artefacts from one job to another, or
to programmatically retrieve artefacts, and artefacts are not visible on
the web UI. Moreover, the artefacts are not versioned, and only the
"latest" is available (and there are concurrency issues when two or more
workflows run at the same time, e.g. on different branches or matrix
jobs).

Finally, only the python workflows tried to upload artefacts; that is
not the case for the rust workflows.

As of mid 2024 (circa end of August), actions/upload-artifact@v2 was
decommissioned, causing python workflows to fail.

Since this was always doubtfully useful, drop the action in all python
workflows.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>