python/vehicle: add support for MQTT over TLS
We are reading the configuration from a .cfg file, where there is no
typing information. So, when we convert it to a dict(), we need to be
careful about the types we need.

For the 'tls' field, this is a little bit delicate; indeed, a non-empty
string evaluates to True, so we can't just convert that field with the
usual cast, bool(val), like we do for integers or floats.

We re-organise the code slightly:
  - rename the variable holding the loaded configuration, so that we
    can re-use it later;
  - move the definition of _set_default() later, to have all
    configuration-parsing code together.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>