python/iqm: don't remove neighbours on http download failure
If the http download fails, we want to keep the current set of
neighbours, if any, and the code already acknowledges that with a
comment.

However, the update callback is always called, even in case of a
download failure, which would cause an empty set of neighbours to
be used, thus actually removing known neighbours, exactly the
opposite of what we wanted.

Fix that by only updating the neighbours when there was no issue
during the download (and parsing).

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>