python/iqm: use only UNIX socket to connect to local MQTT broker
its-iqm only ever connects to a broker that *is* local to the
instance (i.e. either in the same device, like an OBU, the same
container, the same docker composition, the same k8s pod, or
whatever), so we expect to be able to connect via a UNIX socket.

socket-path is now a required configuration item, so do not provide
a default for it.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>